Trello Epic Rollups and separate helper modules
Add trello helper support for rolling up cards from team boards into a central
epic board.

Also, separate out the trello helper into bug and rollup modules.
